Some unit testing tweaks
