+++ test/nools/lib.tasks.spec.js
@@ -1,0 +1,230 @@
+const chai = require('chai');
+const assert = chai.assert;
+chai.use(require('chai-shallow-deep-equal'));
+const {
+  TEST_DAY,
+  reset,
+  aReportBasedTask,
+  aPersonBasedTask,
+  aPlaceBasedTask,
+  aScheduledTaskBasedTask,
+  aReport,
+  aReportWithScheduledTasks,
+  personWithoutReports,
+  personWithReports,
+  placeWithoutReports,
+  loadLibWith,
+} = require('./mocks');
+
+describe('nools lib', function() {
+  beforeEach(() => reset());
+
+  describe('tasks', function() {
+    describe('person-based', function() {
+
+      it('should emit once for a person based task', function() {
+        // given
+        const config = {
+          c: personWithoutReports(),
+          targets: [],
+          tasks: [ aPersonBasedTask() ],
+        };
+
+        // when
+        const emitted = loadLibWith(config).emitted;
+
+        // then
+        assert.shallowDeepEqual(emitted, [
+          {
+            _type: 'task',
+            date: TEST_DAY,
+            resolved: false,
+            actions:[ { form:'example-form' } ]
+          },
+        ]);
+      });
+    });
+
+    describe('place-based', function() {
+
+      it('should emit once for a place based task', function() {
+        // given
+        const config = {
+          c: placeWithoutReports(),
+          targets: [],
+          tasks: [ aPlaceBasedTask() ],
+        };
+
+        // when
+        const emitted = loadLibWith(config).emitted;
+
+        // then
+        assert.shallowDeepEqual(emitted, [
+          { _type:'task', date:TEST_DAY },
+          { _type:'_complete', _id:true },
+        ]);
+      });
+    });
+
+    describe('report-based', function() {
+
+      it('should not emit if contact has no reports', function() {
+        // given
+        const config = {
+          c: personWithoutReports(),
+          targets: [],
+          tasks: [ aReportBasedTask() ],
+        };
+
+        // when
+        const emitted = loadLibWith(config).emitted;
+
+        // then
+        assert.deepEqual(emitted, [
+          { _type:'_complete', _id:true },
+        ]);
+      });
+
+      it('should emit once for a single report', function() {
+        // given
+        const config = {
+          c: personWithReports(aReport()),
+          targets: [],
+          tasks: [ aReportBasedTask() ],
+        };
+
+        // when
+        const emitted = loadLibWith(config).emitted;
+
+        // then
+        assert.shallowDeepEqual(emitted, [
+          { _type:'task', date:TEST_DAY },
+          { _type:'_complete', _id:true },
+        ]);
+      });
+
+      it('should emit once per report', function() {
+        // given
+        const config = {
+          c: personWithReports(aReport(), aReport(), aReport()),
+          targets: [],
+          tasks: [ aReportBasedTask() ],
+        };
+
+        // when
+        const emitted = loadLibWith(config).emitted;
+
+        // then
+        assert.shallowDeepEqual(emitted, [
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'_complete', _id:true },
+        ]);
+      });
+
+      it('should emit once per report per task', function() {
+        // given
+        const config = {
+          c: personWithReports(aReport(), aReport(), aReport()),
+          targets: [],
+          tasks: [ aReportBasedTask(), aReportBasedTask() ],
+        };
+
+        // when
+        const emitted = loadLibWith(config).emitted;
+
+        // then
+        assert.shallowDeepEqual(emitted, [
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'_complete', _id:true },
+        ]);
+      });
+
+      it('should allow custom action content', function() {
+        // given
+        const task = aReportBasedTask();
+        task.actions[0].modifyContent =
+            (r, content) => { content.report_id = r._id; };
+        // and
+        const config = {
+          c: personWithReports(aReport()),
+          targets: [],
+          tasks: [ task ],
+        };
+
+        // when
+        const emitted = loadLibWith(config).emitted;
+
+        // then
+        assert.shallowDeepEqual(emitted, [
+          {
+            actions:[
+              {
+                type: 'report',
+                form: 'example-form',
+                label: 'Follow up',
+                content: {
+                  source: 'task',
+                  source_id: 'r-2',
+                  contact: {
+                    _id: 'c-3',
+                  },
+                  report_id: 'r-2',
+                },
+              },
+            ]
+          },
+        ]);
+      });
+
+    });
+
+    describe('scheduled-task based', function() {
+      it('???', function() { // FIXME this test needs a proper name
+        // given
+        const config = {
+          c: personWithReports(aReportWithScheduledTasks(5)),
+          targets: [],
+          tasks: [ aScheduledTaskBasedTask() ],
+        };
+
+        // when
+        const emitted = loadLibWith(config).emitted;
+
+        // then
+        assert.shallowDeepEqual(emitted, [
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'task', date:TEST_DAY },
+          { _type:'_complete', _id:true },
+        ]);
+      });
+    });
+
+    describe('invalid task type', function() {
+      it('should throw error', function() {
+        // given
+        const invalidTask = aScheduledTaskBasedTask();
+        invalidTask.appliesTo = 'unknown';
+        const config = {
+          c: personWithReports(aReportWithScheduledTasks(5)),
+          targets: [],
+          tasks: [ invalidTask ],
+        };
+
+        // should throw error
+        assert.throws(function() { loadLibWith(config); }, Error,
+          "Error: unrecognised task type: unknown");
+      });
+    });
+
+  });
+});
